feat(dilesa-ventas): pagaré firmado obligatorio en fase 8 (se mueve de la 10)
ADR-048 ya había movido la *definición* del crédito directo a la fase 8
(Dictaminar), pero el *documento del pagaré firmado* (slot rol `pagare`) había
quedado como opcional en la fase 10 — un cabo suelto que Beto detectó.

- Fase 8: slot "Pagaré firmado" (requerido) tras el bloque de crédito directo,
  en el form de cierre y en el de "ya cerrada" (para ventas ya dictaminadas);
  gate en onSubmit que impide cerrar la fase sin el pagaré cuando hay crédito
  directo (coberturaGastos.pagareNecesario > 0).
- Fase 10: se elimina el slot del pagaré y todas sus referencias (queda solo
  fecha/hora de firma + Póliza de Garantía).
- ADR-048: D7 registra la decisión (Beto 2026-06-24).
